capitulo_5/recursos/ejercicio_1/date_producer/requirements.txt
kafka-python==1.4.7
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
